+++ packages/plugin-gallery/web/lib/convert-item-data.js
@@ -1,5 +1,5 @@
 /* eslint-disable complexity */
-import { normalizeUrl } from 'wix-rich-content-common';
+import { normalizeUrl, uuid } from 'wix-rich-content-common';
 import { getAbsoluteUrl } from './baseUrlConverter';
 
 /**
@@ -84,6 +84,8 @@ export const convertItemData = ({ items, anchorTarget, relValue }) =>
       ...item,
       metadata: undefined,
       ...convertedData,
-      itemId: item.url + '_' + index.toString(),
+      // uuid is used to guarentee there are no 2 gallery items on the DOM with the same ID (on different galleries) due to a pro gallery issues with videos - it gives console errors and might also break the video controls behavoiur.
+      // TODO: remove uuid when gallery fixes the issue \ release version 4.
+      itemId: uuid() + '_' + index.toString(),
     };
   });