+++ packages/plugin-gallery/web/src/gallery-viewer.jsx
@@ -7,7 +7,7 @@ import { isEqual, debounce } from 'lodash';
 import { convertItemData } from '../lib/convert-item-data';
 import { getFixedStyles, isHorizontalLayout, sampleItems } from './defaults';
 import { resizeMediaUrl } from '../lib/resize-media-url';
-import { GALLERY_LAYOUTS } from '../lib/layout-data-provider';
+import { GALLERY_LAYOUTS } from './layout-data-provider';
 import styles from '../statics/styles/viewer.rtlignore.scss';
 import '../statics/styles/gallery-styles.rtlignore.scss';
 import ExpandIcon from './icons/expand';
@@ -117,11 +117,8 @@ class GalleryViewer extends React.Component {
 
   getDimensions = () => {
     const width = Math.floor(this.containerRef.current.getBoundingClientRect().width);
-    if (isHorizontalLayout(this.props.componentData.styles)) {
-      const height = this.getGalleryHeight(width);
-      return { width, height };
-    }
-    return { width };
+    const height = this.getGalleryHeight(width);
+    return { width, height };
   };
 
   updateDimensions = debounce(() => {
@@ -165,7 +162,7 @@ class GalleryViewer extends React.Component {
         }
         break;
       case GALLERY_EVENTS.ITEM_ACTION_TRIGGERED:
-        !data.linkData.url && this.hasExpand() && this.handleExpand(data);
+        data.type !== 'video' && !data.linkData.url && this.hasExpand() && this.handleExpand(data);
         break;
       default:
         break;
@@ -178,7 +175,7 @@ class GalleryViewer extends React.Component {
       helpers = {},
     } = this.props;
     helpers.onViewerAction?.(GALLERY_TYPE, 'Click', 'expand_gallery');
-    this.hasExpand() && onExpand?.(this.props.blockKey, data.idx);
+    this.hasExpand() && onExpand(this.props.blockKey, data.idx);
   };
 
   renderExpandIcon = itemProps => {
@@ -188,6 +185,7 @@ class GalleryViewer extends React.Component {
           className={this.styles.expandIcon}
           onClick={e => {
             e.preventDefault();
+            e.stopPropagation();
             this.handleExpand(itemProps);
           }}
         />