+++ packages/plugin-image/web/src/editor.ts
@@ -8,13 +8,15 @@ import { createRicosExtensions } from './tiptap';
 import { TiptapEditorPlugin } from 'wix-tiptap-editor';
 
 export const pluginImage: EditorPluginCreator<ImagePluginEditorConfig> = config => {
-  const pluginConfig = { ...DEFAULTS.config, ...config };
+  const pluginConfig: ImagePluginEditorConfig = { ...DEFAULTS.config, ...config };
   return {
     config: pluginConfig,
     type: IMAGE_TYPE,
     createPlugin: createImagePlugin,
     ModalsMap,
     createPluginData: createImageData,
+    configFixer: ({ helpers }) =>
+      (pluginConfig.uploadHandler = helpers?.handleFileSelection || helpers?.handleFileUpload),
     tiptapExtensions: createRicosExtensions(pluginConfig),
   } as TiptapEditorPlugin;
 };