+++ packages/ricos-content/web/lib/createEmptyContent.ts
@@ -1,0 +1,1 @@
+export { createEmptyContent } from '../src/RicosContentAPI/createEmptyContent';